+++ libatalk/asp/asp_write.c
@@ -1,5 +1,5 @@
 /*
- * $Id: asp_write.c,v 1.3 2001-06-29 14:14:46 rufustfirefly Exp $
+ * $Id: asp_write.c,v 1.4 2009-10-22 12:35:39 franklahm Exp $
  *
  * Copyright (c) 1990,1991 Regents of The University of Michigan.
  * All Rights Reserved.
@@ -39,7 +39,7 @@
 #define memmove(a, b, n)   bcopy((b), (a), (n))
 #endif /* BSD || BSD4_3 */
 
-int asp_wrtcont(ASP asp, char *buf, int *buflen)
+int asp_wrtcont(ASP asp, char *buf, size_t *buflen)
 {
     struct iovec	iov[ ASP_MAXPACKETS ];
     struct atp_block	atpb;
